Fixed bug 1114 (small Longs could be faster in vapi-conduit)
Longs that can fit header and payload into a single bounce buffer get
packed at the sender and unpacked at the receiver, resulting in significant
savings.  Note the following testam timings for 1-byte payloads:

BEFORE:
      1 AMLong      ping-pong roundtrip ReqRep:  0.028 sec   27.845 us    0.068 MB/s
      1 AMLong      ping-pong roundtrip ReqReq:  0.028 sec   28.295 us    0.067 MB/s
      1 AMLong      flood     one-way   Req   :  0.004 sec    3.873 us    0.246 MB/s
      1 AMLong      flood     roundtrip ReqRep:  0.005 sec    5.170 us    0.369 MB/s

AFTER:
      1 AMLong      ping-pong roundtrip ReqRep:  0.017 sec   17.042 us    0.112 MB/s
      1 AMLong      ping-pong roundtrip ReqReq:  0.017 sec   17.422 us    0.109 MB/s
      1 AMLong      flood     one-way   Req   :  0.003 sec    2.867 us    0.333 MB/s
      1 AMLong      flood     roundtrip ReqRep:  0.004 sec    3.667 us    0.520 MB/s

+++ ibv-conduit/gasnet_core.h
@@ -1,6 +1,6 @@
 /*   $Source: /Users/kamil/work/gasnet-cvs2/gasnet/ibv-conduit/gasnet_core.h,v $
- *     $Date: 2005/05/17 20:42:38 $
- * $Revision: 1.36 $
+ *     $Date: 2005/05/18 05:54:49 $
+ * $Revision: 1.37 $
  * Description: GASNet header for vapi conduit core
  * Copyright 2002, Dan Bonachea <bonachea@cs.berkeley.edu>
  * Terms of use are as specified in license.txt
@@ -168,7 +168,7 @@ void gasnetc_counter_wait(gasnetc_counter_t *counter, int handler_context) {
 #define GASNETC_MAX_ARGS	(GASNETC_MAX_ARGS_USER + GASNETC_MAX_ARGS_EXTRA)
 #define GASNETC_MAX_MEDIUM	\
 		(GASNETC_BUFSZ - GASNETC_ALIGNUP(GASNETC_MEDIUM_HDRSZ + 4*GASNETC_MAX_ARGS, 8))
-#define GASNETC_MAX_LONG_REQ	((size_t)0x7fffffff)
+#define GASNETC_MAX_LONG_REQ	(0x7fffffff)
 #if GASNETC_PIN_SEGMENT
   #define GASNETC_MAX_LONG_REP  GASNETC_MAX_LONG_REQ
 #else
